The big one: Initialize Yesod and models
This first commit has a lot of the barebones basics set up for the
fitness-app's backend model. Future changes will be more granularly
recorded in the commit history, but out of laziness I didn't want to
have a bunch of commits while cleaning out the yesod scaffold and
fleshing out the data prototype.