Create Homepage prototype. Document Home handlers.
This prototype is currently functional from backend to frontend, though it is
ugly as sin (of course). Adding a <script> tag to the appropriate stage will be
all that's necessary to bring us right into the JS App from this prototype, so
we're basically ready to go at this point.